+++ src/util/strings.c
@@ -11,12 +11,12 @@
 #include "minishell.h"
 #include "my_string.h"
 
-char *get_next_argument(char *str, int index)
+char *get_next_argument(char *str, size_t index)
 {
     bool text = false;
     bool out = false;
-    int size = strlen(str);
-    int i = index;
+    size_t size = strlen(str);
+    size_t i = index;
 
     for (i = index; i < size && !out; i++) {
         if (str[i] != ' ' && str[i] != '\t') {
@@ -24,12 +24,12 @@ char *get_next_argument(char *str, int index)
         }
         out = ((str[i] == ' ' || str[i] == '\t') && text);
     }
-    return (my_substr_size(str, index, index + i, size));
+    return (my_substr_size(str, (int) index, (int) (index + i), (int) size));
 }
 
 bool is_command_empty(command_t *command)
 {
-    int input = strlen(command->input);
+    size_t input = strlen(command->input);
     char *first = command->args[0];
 
     return (input == 0 || first == NULL || strlen(first) == 0);
